[1230] Add dynamic publish back link helper
