+++ app/views/publish/providers/accredited_provider_search/new.html.erb
@@ -1,14 +1,14 @@
 <% content_for :page_title, title_with_error_prefix("Enter a provider name, UKPRN or postcode", @accredited_provider_search_form.errors.any?) %>
 
-<% content_for :before_content do %>
-  <%= govuk_back_link_to(publish_provider_recruitment_cycle_accredited_providers_path(@provider.provider_code, @provider.recruitment_cycle_year)) %>
-<% end %>
-
 <%= form_with(
         model: @accredited_provider_search_form,
         url: search_publish_provider_recruitment_cycle_accredited_providers_path,
         method: :post
       ) do |f| %>
+
+<% content_for :before_content do %>
+  <%= govuk_back_link_to(publish_back_link_for_adding_accrediting_provider_path(param_form_key: f.object_name.to_sym, params:, recruitment_cycle_year: @recruitment_cycle.year, provider: @provider)) %>
+<% end %>
 
   <%= f.govuk_error_summary %>
