+++ app/views/publish/providers/accredited_providers/new.html.erb
@@ -12,7 +12,7 @@
     <div class="govuk-grid-column-two-thirds">
 
   <% content_for :before_content do %>
-    <%= govuk_back_link_to(search_publish_provider_recruitment_cycle_accredited_providers_path(provider_code: @provider.provider_code)) %>
+    <%= govuk_back_link_to(publish_back_link_for_adding_accrediting_provider_path(param_form_key: f.object_name.to_sym, params:, recruitment_cycle_year: @recruitment_cycle.year, provider: @provider)) %>
   <% end %>
 
         <%= f.govuk_text_area(
@@ -24,7 +24,7 @@
           rows: 10
         ) %>
 
-        <%= f.hidden_field :accredited_provider_id, value: accredited_provider_id %>
+        <%= f.hidden_field :accredited_provider_id, value: params[:accredited_provider_id] || @accredited_provider_form.accredited_provider_id %>
 
         <%= f.govuk_submit t("continue") %>
